Now allow left and right modifier keys to be held down independently, this allows a maximum buy/sell of 5*5*20*20*500*500 = 2.5B! Also modified Format::Number to show "K" for numbers in the thousands, and also reduced the number of digits behind the decimal to 2 to allow large numbers to fit within the game display areas.

+++ source/OutfitterPanel.cpp
@@ -241,8 +241,8 @@ bool OutfitterPanel::CanBuy() const
 
 void OutfitterPanel::Buy()
 {
-	int modifier = Modifier();
-	for(int i = 0; i < modifier && CanBuy(); ++i)
+	int64_t modifier = Modifier();
+	for(int64_t i = 0; i < modifier && CanBuy(); ++i)
 	{
 		// Special case: maps.
 		int mapSize = selectedOutfit->Get("map");
